Removed Lua 5.1 compiler, fixed Print a bit
