+++ src/main/java/org/squiddev/cobalt/Print52.java
@@ -44,48 +44,6 @@ public class Print52 {
 	public static final String[] OPNAMES = {
 		"MOVE",
 		"LOADK",
-		"LOADBOOL",
-		"LOADNIL",
-		"GETUPVAL",
-		"GETGLOBAL",
-		"GETTABLE",
-		"SETGLOBAL",
-		"SETUPVAL",
-		"SETTABLE",
-		"NEWTABLE",
-		"SELF",
-		"ADD",
-		"SUB",
-		"MUL",
-		"DIV",
-		"MOD",
-		"POW",
-		"UNM",
-		"NOT",
-		"LEN",
-		"CONCAT",
-		"JMP",
-		"EQ",
-		"LT",
-		"LE",
-		"TEST",
-		"TESTSET",
-		"CALL",
-		"TAILCALL",
-		"RETURN",
-		"FORLOOP",
-		"FORPREP",
-		"TFORLOOP",
-		"SETLIST",
-		"CLOSE",
-		"CLOSURE",
-		"VARARG",
-		null,
-	};
-
-	public static final String[] OPNAMES_52 = {
-		"MOVE",
-		"LOADK",
 		"LOADKX",
 		"LOADBOOL",
 		"LOADNIL",
@@ -190,20 +148,6 @@ public class Print52 {
 	}
 
 	/**
-	 * Print the code in a prototype
-	 *
-	 * @param f the {@link Prototype}
-	 */
-	public static void printCode(PrintStream ps, Prototype f) {
-		int[] code = f.code;
-		int pc, n = code.length;
-		for (pc = 0; pc < n; pc++) {
-			printOpcode(ps, f, pc);
-			ps.println();
-		}
-	}
-
-	/**
 	 * Print an opcode in a prototype
 	 *
 	 * @param ps the {@link PrintStream} to print to
@@ -226,7 +170,7 @@ public class Print52 {
 		} else {
 			ps.print("[-]  ");
 		}
-		ps.print((f.isLua52 ? OPNAMES_52[o] : OPNAMES[o]) + "  ");
+		ps.print(OPNAMES[o] + "  ");
 		switch (getOpMode(o)) {
 			case iABC:
 				ps.print(a);
@@ -257,6 +201,10 @@ public class Print52 {
 				ps.print("  ; ");
 				printConstant(ps, f, bx);
 				break;
+			case OP_LOADKX:
+				ps.print("  ; ");
+				printConstant(ps, f, GETARG_Ax(code[pc+1]));
+				break;
 			case OP_GETUPVAL:
 			case OP_SETUPVAL:
 				ps.print("  ; ");
@@ -266,10 +214,17 @@ public class Print52 {
 					ps.print("-");
 				}
 				break;
-			case OP_GETGLOBAL:
-			case OP_SETGLOBAL:
+			case OP_GETTABUP:
+			case OP_SETTABUP:
 				ps.print("  ; ");
-				printConstant(ps, f, bx);
+				if (f.upvalues.length > b) {
+					printValue(ps, f.upvalues[b]);
+				} else {
+					ps.print("-");
+				}
+				ps.print("[");
+				printConstant(ps, f, c);
+				ps.print("]");
 				break;
 			case OP_GETTABLE:
 			case OP_SELF:
@@ -305,6 +260,7 @@ public class Print52 {
 			case OP_JMP:
 			case OP_FORLOOP:
 			case OP_FORPREP:
+			case OP_TFORLOOP:
 				ps.print("  ; to " + (sbx + pc + 2));
 				break;
 			case OP_CLOSURE:
@@ -327,151 +283,5 @@ public class Print52 {
 
 	private static int getline(Prototype f, int pc) {
 		return pc > 0 && f.lineinfo != null && pc < f.lineinfo.length ? f.lineinfo[pc] : -1;
-	}
-
-	private static void printHeader(PrintStream ps, Prototype f) {
-		String s = String.valueOf(f.source);
-		if (s.startsWith("@") || s.startsWith("=")) {
-			s = s.substring(1);
-		} else if ("\033Lua".equals(s)) {
-			s = "(bstring)";
-		} else {
-			s = "(string)";
-		}
-		String a = (f.linedefined == 0) ? "main" : "function";
-		ps.print("\n%" + a + " <" + s + ":" + f.linedefined + ","
-			+ f.lastlinedefined + "> (" + f.code.length + " instructions, "
-			+ f.code.length * 4 + " bytes at " + id(f) + ")\n");
-		ps.print(f.numparams + " param, " + f.maxstacksize + " slot, "
-			+ f.upvalues.length + " upvalue, ");
-		ps.print(f.locvars.length + " local, " + f.k.length
-			+ " constant, " + f.p.length + " function\n");
-	}
-
-	private static void printConstants(PrintStream ps, Prototype f) {
-		int i, n = f.k.length;
-		ps.print("constants (" + n + ") for " + id(f) + ":\n");
-		for (i = 0; i < n; i++) {
-			ps.print("  " + (i + 1) + "  ");
-			printValue(ps, f.k[i]);
-			ps.print("\n");
-		}
-	}
-
-	private static void printLocals(PrintStream ps, Prototype f) {
-		int i, n = f.locvars.length;
-		ps.print("locals (" + n + ") for " + id(f) + ":\n");
-		for (i = 0; i < n; i++) {
-			ps.println("  " + i + "  " + f.locvars[i].name + " " + (f.locvars[i].startpc + 1) + " " + (f.locvars[i].endpc + 1));
-		}
-	}
-
-	private static void printUpValues(PrintStream ps, Prototype f) {
-		int i, n = f.upvalues.length;
-		ps.print("upvalues (" + n + ") for " + id(f) + ":\n");
-		for (i = 0; i < n; i++) {
-			ps.print("  " + i + "  " + f.upvalues[i] + "\n");
-		}
-	}
-
-	public static String show(Prototype p) {
-		return showWith(ps -> printFunction(ps, p, true));
-	}
-
-	public static void printFunction(PrintStream ps, Prototype f, boolean full) {
-		int i, n = f.p.length;
-		printHeader(ps, f);
-		printCode(ps, f);
-		if (full) {
-			printConstants(ps, f);
-			printLocals(ps, f);
-			printUpValues(ps, f);
-		}
-		for (i = 0; i < n; i++) {
-			printFunction(ps, f.p[i], full);
-		}
-	}
-
-	private static void format(PrintStream ps, String s, int maxcols) {
-		int n = s.length();
-		if (n > maxcols) {
-			ps.print(s.substring(0, maxcols));
-		} else {
-			ps.print(s);
-			for (int i = maxcols - n; --i >= 0; ) {
-				ps.print(' ');
-			}
-		}
-	}
-
-	private static String id(Prototype f) {
-		return f.sourceShort() + ":" + f.linedefined;
-	}
-
-	/**
-	 * Print the state of a {@link LuaClosure} that is being executed
-	 *
-	 * @param cl      the {@link LuaClosure}
-	 * @param pc      the program counter
-	 * @param stack   the stack of {@link LuaValue}
-	 * @param top     the top of the stack
-	 * @param varargs any {@link Varargs} value that may apply
-	 */
-	public static void printState(PrintStream ps, LuaClosure cl, int pc, LuaValue[] stack, int top, Varargs varargs) {
-		// print opcode into buffer
-		format(ps, showWith(p -> printOpcode(p, cl.getPrototype(), pc)), 50);
-
-		// print stack
-		ps.print('[');
-		for (int i = 0; i < stack.length; i++) {
-			LuaValue v = stack[i];
-			if (v == null) {
-				ps.print("null");
-			} else {
-				switch (v.type()) {
-					case Constants.TSTRING:
-						LuaString s = (LuaString) v;
-						ps.print(s.length() < 48 ?
-							s.toString() :
-							s.substring(0, 32).toString() + "...+" + (s.length() - 32) + "b");
-						break;
-					case Constants.TFUNCTION:
-						ps.print((v instanceof LuaClosure) ?
-							((LuaClosure) v).getPrototype().toString() : v.toString());
-						break;
-					case Constants.TUSERDATA:
-						Object o = v.toUserdata();
-						if (o != null) {
-							String n = o.getClass().getName();
-							n = n.substring(n.lastIndexOf('.') + 1);
-							ps.print(n + ": " + Integer.toHexString(o.hashCode()));
-						} else {
-							ps.print(v.toString());
-						}
-						break;
-					default:
-						ps.print(v.toString());
-				}
-			}
-			if (i + 1 == top) {
-				ps.print(']');
-			}
-			ps.print(" | ");
-		}
-		ps.print(varargs);
-		ps.println();
-	}
-
-	private static String showWith(Consumer<PrintStream> f) {
-		ByteArrayOutputStream output = new ByteArrayOutputStream();
-		PrintStream ps;
-		try {
-			ps = new PrintStream(output, false, "UTF-8");
-		} catch (UnsupportedEncodingException e) {
-			throw new IllegalStateException("UTF-8 should exist", e);
-		}
-		f.accept(ps);
-		ps.flush();
-		return new String(output.toByteArray(), StandardCharsets.UTF_8);
 	}
 }
